Backend/app/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
motor
pydantic
email-validator